requirements.txt
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
configparser==5.0.0
cycler==0.10.0
google-api-core==1.22.2
google-auth==1.21.2
google-cloud-language==1.3.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
instaloader==4.5.3
kiwisolver==1.2.0
minio==6.0.0
numpy==1.19.2
peewee==3.13.3
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyJWT==1.7.1
pyparsing==2.4.7
python-arango==6.0.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
toml==0.10.1
urllib3==1.25.10
